+++ southwest.py
@@ -15,6 +15,8 @@ Log into your account:
 
 Options:
     --test-notifications Test the notification URLs configuration and exit
+    --debug-screenshots  Take screenshots of the browser for debugging purposes. Screenshots
+                         will be stored in the 'logs/' directory
     -v, --verbose        Display debug messages
     -h, --help           Display this help and exit
     -V, --version        Display version information and exit